Create reactive chains, add localStorage integration, add circular serialization interface, optimize graph and node constructors, begin full API development
Now there's a Chain constructor that is used to cache and react to
asynchronously evaluated chain contexts (no instance caching yet). Any
methods that change the chain context will generate a new chain instance
for the subsequent chains to interact with.

The localStorage implementation can now listen for read requests and
feed the values back through into the calling graph.

When the graph constructor is given an object parameter, it'll serialize it into a graph.

The node and graph constructors have been optimized for performance.

With the new Chain constructor, the API is now ready to be developed
fully (let's hope). At this point, the only methods built are `.get` and
`.put`.