Fix infinite merge recursion on graph watch, add tests for Graph & Node
Due to issues with conflict resolution, merges from primitives would always be an update and HAM wouldn't even check for state. It's been updated to always try to resolve the state, otherwise treat it as a brand new update. When reading back from localStorage to do a merge into the current state, it would trigger an event update, which would trigger a `graph.watch` change event, causing localStorage to attempt another merge, and so forth.

Tests have been added to guard against regression and easily verify quality of refactors.