Remove old deps, allow lex mutation, chain improvements, clean up code
Old dependencies and code have been removed.

Lexes now have a method called ".set" that changes a property and wipes the ID to be re-evaluated later. This allows you to clone lexes, mutate them, and then re-evaluate the ID.

The chain now has a method to iterate up the chain until it reaches root or returns true. Also, it no longer tries to diff the chain and leaves that up to a higher level to handle. It's now scoping the events, and allows storage of scope variables.