Remove homepage, index.js, and shoot/, make entry file isomorphic for tests, fix broken refs, fix exports.
The entry file had been exporting an object containing gun, chain, lex and others and then merging with the window. Instead, it now exports the Gun method and checks to see if window is defined before trying to merge with it.

The gunDB websites have their own repos on github. I don't see a reason to keep the index.html file there anymore. Also, since `package.json` points to the entry file, there's no reason to keep the index.js file at root level.

There was a remnant of a localStorage experiment were causing a crash on node, so it's been removed.