Remove caret semver operator (npm compatibility issues)
So it seems that travis-ci, when nvm-installing really old versions of node, doesn't support the caret version operator. For that reason, I've removed it and locked in versions. In the future I hope to loosen up the deps for more fault-tolerant installs, but in the meantime I'll bet EventEmitter3 and Mocha have enough redundant cdn coverage to support a strict version.